+++ BZI/pseudopots.py
@@ -13,131 +13,6 @@ angstrom_to_Bohr = 1.889725989
 Ry_to_eV = 13.605698066
 isclose = np.vectorize(np.isclose)
 
-# class EmpiricalPP(object):
-#     """Create an empirical pseudopotential.
-
-#     Args:
-#         lattice (:py:obj:`BZI.symmetry.lattice`): an instance of Lattice.
-#         form_factors (list): a list of pseudopotential form factors. Every 
-#             energy shell up to the cutoff energy should be accounted for.
-#         energy_cutoff (float): the cutoff energy of the Fourier expansion.  
-#         atomic_positions (list): a list of atomic positions.
-#         nvalence_electrons (int): the number of valence electrons
-#         energy_shift (float): an energy shift typically used to place the Fermi
-#             level at the correct position.
-#         fermi_level (float): the fermi level.
-#         total_energy (float): the total energy.
-
-#     Attributes:
-#         lattice (:py:obj:`BZI.symmetry.lattice`): an instance of Lattice.
-#         form_factors (list): a list of pseudopotential form factors. Every 
-#             energy shell up to the cutoff energy should be accounted for.
-#         energy_cutoff (float): the cutoff energy of the Fourier expansion.  
-#         rlat_pts (list): a list of reciprocal lattice points included in the 
-#             Fourier expansion.
-#         energy_shells (list): a list of spherical shells that points in rlat_pts
-#             reside on.
-#         atomic_positions (list): a list of atomic positions.
-#         nvalence_electrons (int): the number of valence electrons
-#         energy_shift (float): an energy shift typically used to place the Fermi
-#             level at the correct position.
-#         total_energy (float): the total energy.
-
-#     Example:
-#         >>> centering_type = "face"
-#         >>> lat_const = 4.0
-#         >>> lat_consts = [lat_const]*3
-#         >>> lat_angles = [np.pi/2]*3
-#         >>> lattice = Lattice(centering_type, lat_consts, lat_angles)
-#         >>> pff = [0.019, 0.055]
-#         >>> cutoff = 4*(2*np.pi/lat_const)**2
-#         >>> atomic_positions = [0.,0.,0.]
-#         >>> Nvalence_electrons = 3
-#         >>> PP = EmpiricalPP(lattice, pff, cutoff, nvalence_electrons,
-#         >>>                  atomic_positions)
-#     """
-    
-#     def __init__(self, lattice, form_factors, energy_cutoff, atomic_positions,
-#                  nvalence_electrons, energy_shift=None, fermi_level=None):
-#         self.lattice = lattice
-#         self.form_factors = form_factors
-#         self.energy_cutoff = energy_cutoff
-#         if np.shape(atomic_positions) == (3,):
-#             msg = ("Please provide a list of atomic positions instead of an "
-#                    "individual atomic position.")
-#             raise ValueError(msg.format(atomic_positions))
-#         else:
-#             self.atomic_positions = atomic_positions
-#         self.rlat_pts = sphere_pts(self.lattice.reciprocal_vectors,
-#                                    self.energy_cutoff)
-#         self.find_energy_shells()
-#         self.nvalence_electrons = nvalence_electrons
-#         self.energy_shift = energy_shift or 0.
-#         self.fermi_level = fermi_level or 0.
-
-#     def find_energy_shells(self):
-#         """Find the spherical shells of constant energy on which the points in
-#         rlat_pts reside. These are ordered from least to greatest.
-#         """
-
-#         shells = []
-#         for rpt in self.rlat_pts:
-#             if any(np.isclose(np.dot(rpt, rpt), shells)):
-#                 continue
-#             else:
-#                 shells.append(np.dot(rpt, rpt))
-#         # Round energy shells at the 10th decimal so that equality comparisons
-#         # can be performed. This could potential cause problems.
-#         self.energy_shells = np.round(np.sort(shells), 10)
-                
-#     def eval(self, kpoint, neigvals):
-#         """Evaluate the empirical pseudopotential eigenvalues at the provided
-#         k-point. Only return the lowest 'neigvals' eigenvalues.
-#         """
-
-#         size = len(self.rlat_pts)
-#         H = np.zeros([size, size], dtype=complex)
-#         for i,rpt1 in enumerate(self.rlat_pts):
-#             for j in range(i+1):
-#                 rpt2 = self.rlat_pts[j]
-#                 h = rpt2 - rpt1
-#                 r2 = np.dot(h,h)
-#                 for ap in self.atomic_positions:
-#                     if i == j:
-#                         H[i,j] = np.dot(kpoint + rpt1, kpoint + rpt1)
-#                         break
-#                     else:
-#                         try:
-#                             k = np.where(np.isclose(self.energy_shells, r2))[0][0]
-#                             H[i,j] += self.form_factors[k]*np.exp(-1j*np.dot(h,ap))
-#                         except:
-#                             continue
-#         return np.sort(np.linalg.eigvalsh(H))[:neigvals]*Ry_to_eV
-
-#     def hamiltonian(self, kpoint):
-#         """Evaluate the empirical pseudopotential Hamiltonian at the provided
-#         k-point. This function is typically used to verify the Hamiltonian is 
-#         Hermitian.
-#         """
-        
-#         size = len(self.rlat_pts)
-#         H = np.zeros([size, size], dtype=complex)
-#         for i,rpt1 in enumerate(self.rlat_pts):
-#             for j,rpt2 in enumerate(self.rlat_pts):
-#                 h = rpt2 - rpt1
-#                 r2 = np.dot(h,h)
-#                 for ap in self.atomic_positions:
-#                     if i == j:
-#                         H[i,j] = np.dot(kpoint + rpt1, kpoint + rpt1)
-#                         break
-#                     else:
-#                         try:
-#                             k = np.where(np.isclose(self.energy_shells, r2))[0][0]
-#                             H[i,j] = self.form_factors[k]*np.exp(-1j*np.dot(h,ap))
-#                         except:
-#                             continue
-#         return H*Ry_to_eV
-
 class EmpiricalPP(object):
     """Create an empirical pseudopotential.
 
@@ -150,6 +25,8 @@ class EmpiricalPP(object):
         nvalence_electrons (int): the number of valence electrons
         energy_shift (float): an energy shift typically used to place the Fermi
             level at the correct position.
+        material (str): a string describing the empirical pseudopotential,
+            such as the chemical formula.
         fermi_level (float): the fermi level.
         total_energy (float): the total energy.
 
@@ -164,8 +41,11 @@ class EmpiricalPP(object):
             reside on.
         atomic_positions (list): a list of atomic positions.
         nvalence_electrons (int): the number of valence electrons
+        material (str): a string describing the empirical pseudopotential,
+            such as the chemical formula.
         energy_shift (float): an energy shift typically used to place the Fermi
             level at the correct position.
+        fermi_level (float): the fermi level.
         total_energy (float): the total energy.
 
     Example:
@@ -183,7 +63,9 @@ class EmpiricalPP(object):
     """
     
     def __init__(self, lattice, form_factors, energy_cutoff, atomic_positions,
-                 nvalence_electrons, energy_shift=None, fermi_level=None):
+                 nvalence_electrons, material, energy_shift=None,
+                 fermi_level=None, total_energy=None):
+        self.material = material
         self.lattice = lattice
         self.form_factors = form_factors
         self.energy_cutoff = energy_cutoff
@@ -199,6 +81,7 @@ class EmpiricalPP(object):
         self.nvalence_electrons = nvalence_electrons
         self.energy_shift = energy_shift or 0.
         self.fermi_level = fermi_level or 0.
+        self.total_energy = total_energy or 0.
         self.init_hamiltonian = self.hamiltonian([0.]*3) - np.diag(
             np.diag(self.hamiltonian([0.]*3)))
 
@@ -213,9 +96,8 @@ class EmpiricalPP(object):
                 continue
             else:
                 shells.append(np.dot(rpt, rpt))
-        # Round energy shells at the 10th decimal so that equality comparisons
-        # can be performed. This could potential cause problems.
-        self.energy_shells = np.round(np.sort(shells), 10)
+                
+        self.energy_shells = np.sort(shells)
                 
     def eval(self, kpoint, neigvals):
         """Evaluate the empirical pseudopotential eigenvalues at the provided
@@ -224,12 +106,12 @@ class EmpiricalPP(object):
         
         if np.allclose(self.atomic_positions, [[0.]*3]):
             diag = np.eye(len(self.rlat_pts))*np.apply_along_axis(
-                norm, 1, self.rlat_pts - kpoint)**2
+                norm, 1, self.rlat_pts + kpoint)**2 
             H = self.init_hamiltonian + diag*Ry_to_eV
         else:
             # Calculate the diagonal elements of the Hamiltonian.
             diag = np.eye(len(self.rlat_pts))*np.apply_along_axis(
-                norm, 1, self.rlat_pts - kpoint)**2
+                norm, 1, self.rlat_pts + kpoint)**2
             # For some reason the lattice points need to be double nested for
             # numpy.tile to work.
             nested_rlatpts = np.array([[rlp] for rlp in self.rlat_pts])
@@ -253,7 +135,7 @@ class EmpiricalPP(object):
         
         # Calculate the diagonal elements of the Hamiltonian.
         diag = np.eye(len(self.rlat_pts))*np.apply_along_axis(norm, 1,
-                                                              self.rlat_pts - kpoint)**2
+                                                              self.rlat_pts + kpoint)**2 
         # For some reason the lattice points need to be double nested for
         # numpy.tile to work.
         nested_rlatpts = np.array([[rlp] for rlp in self.rlat_pts])
@@ -285,9 +167,11 @@ class CohenEmpiricalPP(object):
         antisym_form_factors (list): a list of anti-symmetric, pseudopotential
             form factors. Every energy shell up to the cutoff energy should be 
             accounted for.
-        energy_cutoff (float): the cutoff energy of the Fourier expansion.  
+        energy_cutoff (float): the cutoff energy of the Fourier expansion.
         atomic_positions (list): a list of atomic positions.
         nvalence_electrons (int): the number of valence electrons.
+        material (str): a string describing the empirical pseudopotential,
+            such as the chemical formula.
         energy_shift (float): an energy shift typically used to place the Fermi
             level at the correct position.
         fermi_level (float): the fermi level.
@@ -308,6 +192,8 @@ class CohenEmpiricalPP(object):
             reside on.
         atomic_positions (list): a list of atomic positions.
         nvalence_electrons (int): the number of valence electrons.
+        material (str): a string describing the empirical pseudopotential,
+            such as the chemical formula.
         energy_shift (float): an energy shift typically used to place the Fermi
             level at the correct position.
         fermi_level (float): the fermi level.
@@ -316,8 +202,10 @@ class CohenEmpiricalPP(object):
 
     def __init__(self, lattice, sym_form_factors, antisym_form_factors,
                  energy_cutoff, atomic_positions, nvalence_electrons,
-                 energy_shift=None, fermi_level=None):
-        
+                 material, energy_shift=None, fermi_level=None,
+                 total_energy=None):
+
+        self.material = material
         self.lattice = lattice
         self.sym_form_factors = sym_form_factors
         self.antisym_form_factors = antisym_form_factors
@@ -334,6 +222,7 @@ class CohenEmpiricalPP(object):
         self.nvalence_electrons = nvalence_electrons
         self.energy_shift = energy_shift or 0.
         self.fermi_level = fermi_level or 0.
+        self.total_energy = total_energy or 0.
 
     def find_energy_shells(self):
         """Find the spherical shells of constant energy on which the points in
@@ -349,58 +238,97 @@ class CohenEmpiricalPP(object):
         self.energy_shells = np.sort(shells)
                     
     def eval(self, kpoint, neigvals):
-        """Evaluate the empirical pseudopotential eigenvalues at the provided
-        k-point. Only return the lowest 'neigvals' eigenvalues.
+        """Evaluate the empirical pseudopotential Hamiltonian at the provided
+        k-point. This function is typically used to verify the Hamiltonian is 
+        Hermitian.
         """
         
-        size = len(self.rlat_pts)
-        H = np.zeros([size, size], dtype=complex)
-        for i,rpt1 in enumerate(self.rlat_pts):
-            for j in range(i+1):
-                rpt2 = self.rlat_pts[j]
-                h = rpt2 - rpt1
-                r2 = np.dot(h,h)
-                for ap in self.atomic_positions:
-                    if i == j:
-                        H[i,j] = np.dot(kpoint + rpt1, kpoint + rpt1)
-                        break
-                    else:
-                        try:
-                            k = np.where(np.isclose(self.energy_shells, r2))[0][0]
-                            H[i,j] = self.sym_form_factors[k]*(
-                                np.cos(np.dot(h,ap))) + (
-                                    1j*self.antisym_form_factors[k]*(
-                                        np.sin(np.dot(h,ap))))
-                        except:
-                            continue
+        # Calculate the diagonal elements of the Hamiltonian.
+        diag = np.eye(len(self.rlat_pts))*np.apply_along_axis(norm, 1,
+                                                              self.rlat_pts + kpoint)**2
+        # For some reason the lattice points need to be double nested for
+        # numpy.tile to work.
+        nested_rlatpts = np.array([[rlp] for rlp in self.rlat_pts])
+
+        # Create a matrix of reciprocal lattice points where the first lattice
+        # point is repeated acroass the first row, the second lattice point the
+        # second row, and so on.
+        rlatpt_mat = np.tile(nested_rlatpts,(len(self.rlat_pts),1))
+        
+        # Create a matrix of the differences of the lattice points. Each element
+        # is given by a_i - a_j.
+        rlp_diff = rlatpt_mat - np.transpose(rlatpt_mat, (1,0,2))
+
+        # Find the norm squared of the difference.
+        r2_mat = np.apply_along_axis(norm, 2, rlp_diff)**2
+
+        # The symmetric part of the Hamiltonian.
+        sff = np.zeros(np.shape(r2_mat))
+        for i in range(1,len(self.sym_form_factors)):
+            if self.sym_form_factors[i] == 0.:
+                continue
+            else:
+                sff[np.isclose(r2_mat, self.energy_shells[i])] = self.sym_form_factors[i]
+                
+        sff *= np.cos(np.sum(rlp_diff*np.sum(self.atomic_positions, 0), 2))
+
+        # The Anti-symmetric part of the Hamiltonian.
+        asff = np.zeros(np.shape(r2_mat), dtype=complex)
+        for i in range(1,len(self.antisym_form_factors)):
+            if self.antisym_form_factors[i] == 0.:
+                continue
+            else:
+                asff[np.isclose(r2_mat, self.energy_shells[i])] = self.antisym_form_factors[i]
+                
+        asff *= 1j*np.sin(np.sum(rlp_diff*np.sum(self.atomic_positions, 0), 2))
+
+        H = (diag + sff + asff)
         return np.sort(np.linalg.eigvalsh(H))[:neigvals]*Ry_to_eV
+
 
     def hamiltonian(self, kpoint):
         """Evaluate the empirical pseudopotential Hamiltonian at the provided
         k-point. This function is typically used to verify the Hamiltonian is 
         Hermitian.
         """
+        # Calculate the diagonal elements of the Hamiltonian.
+        diag = np.eye(len(self.rlat_pts))*np.apply_along_axis(norm, 1,
+                                                              self.rlat_pts + kpoint)**2
+        # For some reason the lattice points need to be double nested for
+        # numpy.tile to work.
+        nested_rlatpts = np.array([[rlp] for rlp in self.rlat_pts])
+        # Create a matrix of reciprocal lattice points where the first lattice
+        # point is repeated acroass the first row, the second lattice point the
+        # second row, and so on.
+        rlatpt_mat = np.tile(nested_rlatpts,(len(self.rlat_pts),1))
+        # Create a matrix of the differences of the lattice points. Each element
+        # is given by a_i - a_j.
+        rlp_diff = rlatpt_mat - np.transpose(rlatpt_mat, (1,0,2))
+        # Find the norm squared of the difference.
+        r2_mat = np.apply_along_axis(norm, 2, rlp_diff)**2
+
+        # The symmetry part of the Hamiltonian.
+        sff = np.zeros(np.shape(r2_mat))
+        for i in range(1,len(self.sym_form_factors)):
+            if self.sym_form_factors[i] == 0.:
+                continue
+            else:
+                sff[np.isclose(r2_mat, self.energy_shells[i])] = self.sym_form_factors[i]
+                
+        sff *= np.cos(np.sum(rlp_diff*np.sum(self.atomic_positions, 0), 2))
+
+        # The anti-symmetry part of the Hamiltonian.
+        asff = np.zeros(np.shape(r2_mat), dtype=complex)
+        for i in range(1,len(self.antisym_form_factors)):
+            if self.antisym_form_factors[i] == 0.:
+                continue
+            else:
+                asff[np.isclose(r2_mat, self.energy_shells[i])] = self.antisym_form_factors[i]
+                
+        asff *= 1j*np.sin(np.sum(rlp_diff*np.sum(self.atomic_positions, 0), 2))
         
-        size = len(self.rlat_pts)
-        H = np.zeros([size, size], dtype=complex)
-        for i,rpt1 in enumerate(self.rlat_pts):
-            for j,rpt2 in enumerate(self.rlat_pts):
-                h = rpt2 - rpt1
-                r2 = np.dot(h,h)
-                for ap in self.atomic_positions:
-                    if i == j:
-                        H[i,j] = np.dot(kpoint + rpt1, kpoint + rpt1)
-                        break
-                    else:
-                        try:
-                            k = np.where(np.isclose(self.energy_shells, r2))[0][0]
-                            H[i,j] = self.sym_form_factors[k]*(
-                                np.cos(np.dot(h, ap))) + (
-                                    1j*self.antisym_form_factors[k]*
-                                    np.sin(np.dot(h, ap)))
-                        except:
-                            continue
-        return H*Ry_to_eV
+        return (diag + sff + asff)*Ry_to_eV
+
 
 class FreeElectronModel():
     """This is the popular free electron model. In this model the potential is
@@ -541,7 +469,8 @@ Si_energy_cutoff = 11.*(2*np.pi/Si_lat_const)**2
 Si_atomic_positions = [Si_lat_const/8.*np.array([1,1,1])] # one atomic basis vector
 Si_nvalence_electrons = 4
 Si_PP = CohenEmpiricalPP(Si_lattice, Si_spff, Si_apff, Si_energy_cutoff,
-                         Si_atomic_positions, Si_nvalence_electrons)
+                         Si_atomic_positions, Si_nvalence_electrons,
+                         material="Si")
 
 #### Pseudopotential of Ge ####
 Ge_lat_centering = "face"
@@ -554,7 +483,8 @@ Ge_cutoff_energy = 11.*(2*np.pi/Ge_lat_const)**2
 Ge_atomic_positions = [Ge_lat_const/8.*np.array([1,1,1])] # one atomic basis vector
 Ge_nvalence_electrons = 4
 Ge_PP = CohenEmpiricalPP(Ge_lattice, Ge_spff, Ge_apff, Ge_cutoff_energy,
-                         Ge_atomic_positions, Ge_nvalence_electrons)
+                         Ge_atomic_positions, Ge_nvalence_electrons,
+                         material="Ge")
 
 #### Pseudopotential of Sn ####
 cSn_lat_centering = "face"
@@ -567,7 +497,8 @@ cSn_cutoff_energy = 11.*(2*np.pi/cSn_lat_const)**2
 cSn_atomic_positions = [cSn_lat_const/8.*np.array([1,1,1])] # one atomic basis vector
 cSn_nvalence_electrons = 4
 cSn_PP = CohenEmpiricalPP(cSn_lattice, Sn_spff, Sn_apff, cSn_cutoff_energy,
-                          cSn_atomic_positions, cSn_nvalence_electrons)
+                          cSn_atomic_positions, cSn_nvalence_electrons,
+                          material="Sn")
 
 #### Pseudopotential of GaP ####
 GaP_lat_centering = "face"
@@ -580,7 +511,8 @@ GaP_cutoff_energy = 11.*(2*np.pi/GaP_lat_const)**2
 GaP_atomic_positions = [GaP_lat_const/8.*np.array([1,1,1])] # one atomic basis vector
 GaP_nvalence_electrons = 8
 GaP_PP = CohenEmpiricalPP(GaP_lattice, GaP_spff, GaP_apff, GaP_cutoff_energy,
-                          GaP_atomic_positions, GaP_nvalence_electrons)
+                          GaP_atomic_positions, GaP_nvalence_electrons,
+                          material="GaP")
 
 #### Pseudopotential of GaAs ####
 GaAs_lat_centering = "face"
@@ -593,7 +525,8 @@ GaAs_cutoff_energy = 11.*(2*np.pi/GaAs_lat_const)**2
 GaAs_atomic_positions = [GaAs_lat_const/8.*np.array([1,1,1])] # one atomic basis vector
 GaAs_nvalence_electrons = 8
 GaAs_PP = CohenEmpiricalPP(GaAs_lattice, GaAs_spff, GaAs_apff, GaAs_cutoff_energy,
-                           GaAs_atomic_positions, GaAs_nvalence_electrons)
+                           GaAs_atomic_positions, GaAs_nvalence_electrons,
+                           material="GaAs")
 
 #### Pseudopotential of AlSb ####
 AlSb_lat_centering = "face"
@@ -607,7 +540,7 @@ AlSb_atomic_positions = [AlSb_lat_const/8.*np.array([1,1,1])] # one atomic basis
 AlSb_nvalence_electrons = 8
 AlSb_PP = CohenEmpiricalPP(AlSb_lattice, AlSb_spff, AlSb_apff,
                            AlSb_cutoff_energy, AlSb_atomic_positions,
-                           AlSb_nvalence_electrons)
+                           AlSb_nvalence_electrons, material="AlSb")
 
 #### Pseudopotential of InP ####
 InP_lat_centering = "face"
@@ -620,7 +553,8 @@ InP_cutoff_energy = 11.*(2*np.pi/InP_lat_const)**2
 InP_atomic_positions = [InP_lat_const/8.*np.array([1,1,1])] # one atomic basis vector
 InP_nvalence_electrons = 8
 InP_PP = CohenEmpiricalPP(InP_lattice, InP_spff, InP_apff, InP_cutoff_energy,
-                          InP_atomic_positions, InP_nvalence_electrons)
+                          InP_atomic_positions, InP_nvalence_electrons,
+                          material="InP")
 
 #### Pseudopotential of GaSb ####
 GaSb_lat_centering = "face"
@@ -634,7 +568,7 @@ GaSb_atomic_positions = [GaSb_lat_const/8.*np.array([1,1,1])] # one atomic basis
 GaSb_nvalence_electrons = 8
 GaSb_PP = CohenEmpiricalPP(GaSb_lattice, GaSb_spff, GaSb_apff,
                            GaSb_cutoff_energy, GaSb_atomic_positions,
-                           GaSb_nvalence_electrons)
+                           GaSb_nvalence_electrons, material="GaSb")
 
 #### Pseudopotential of InAs ####
 InAs_lat_centering = "face"
@@ -648,7 +582,7 @@ InAs_atomic_positions = [InAs_lat_const/8.*np.array([1,1,1])] # one atomic basis
 InAs_nvalence_electrons = 8
 InAs_PP = CohenEmpiricalPP(InAs_lattice, InAs_spff, InAs_apff,
                            InAs_cutoff_energy, InAs_atomic_positions,
-                           InAs_nvalence_electrons)
+                           InAs_nvalence_electrons, material="InAs")
 
 #### Pseudopotential of InSb ####
 InSb_lat_centering = "face"
@@ -662,7 +596,7 @@ InSb_atomic_positions = [InSb_lat_const/8.*np.array([1,1,1])] # one atomic basis
 InSb_nvalence_electrons = 8
 InSb_PP = CohenEmpiricalPP(InSb_lattice, InSb_spff, InSb_apff,
                            InSb_cutoff_energy, InSb_atomic_positions,
-                           InSb_nvalence_electrons)
+                           InSb_nvalence_electrons, material="InSb")
 
 #### Pseudopotential of ZnS ####
 ZnS_lat_centering = "face"
@@ -676,7 +610,7 @@ ZnS_atomic_positions = [ZnS_lat_const/8.*np.array([1,1,1])] # one atomic basis v
 ZnS_nvalence_electrons = 18
 ZnS_PP = CohenEmpiricalPP(ZnS_lattice, ZnS_spff, ZnS_apff,
                           ZnS_cutoff_energy, ZnS_atomic_positions,
-                          ZnS_nvalence_electrons)
+                          ZnS_nvalence_electrons, material="ZnS")
 
 #### Pseudopotential of ZnSe ####
 ZnSe_lat_centering = "face"
@@ -690,7 +624,7 @@ ZnSe_atomic_positions = [ZnSe_lat_const/8.*np.array([1,1,1])] # one atomic basis
 ZnSe_nvalence_electrons = 18
 ZnSe_PP = CohenEmpiricalPP(ZnSe_lattice, ZnSe_spff, ZnSe_apff,
                            ZnSe_cutoff_energy, ZnSe_atomic_positions,
-                           ZnSe_nvalence_electrons)
+                           ZnSe_nvalence_electrons, material="ZnSe")
 
 #### Pseudopotential of ZnTe ####
 ZnTe_lat_centering = "face"
@@ -704,7 +638,7 @@ ZnTe_atomic_positions = [ZnTe_lat_const/8.*np.array([1,1,1])] # one atomic basis
 ZnTe_nvalence_electrons = 18
 ZnTe_PP = CohenEmpiricalPP(ZnTe_lattice, ZnTe_spff, ZnTe_apff,
                            ZnTe_cutoff_energy, ZnTe_atomic_positions,
-                           ZnTe_nvalence_electrons)
+                           ZnTe_nvalence_electrons, material="ZnTe")
 
 #### Pseudopotential of CdTe ####
 CdTe_lat_centering = "face"
@@ -718,7 +652,7 @@ CdTe_atomic_positions = [CdTe_lat_const/8.*np.array([1,1,1])] # one atomic basis
 CdTe_nvalence_electrons = 18
 CdTe_PP = CohenEmpiricalPP(CdTe_lattice, CdTe_spff, CdTe_apff,
                            CdTe_cutoff_energy, CdTe_atomic_positions,
-                           CdTe_nvalence_electrons)
+                           CdTe_nvalence_electrons, material="CdTe")
 
 # The end of Cohen's 14 pseudopotentials with diamond or zinc-blende structure.
 
@@ -734,7 +668,7 @@ Toy_energy_cutoff = 2*(2*np.pi/Toy_lat_const)**2
 Toy_atomic_positions = [[0.]*3]
 Toy_nvalence_electrons = 3
 Toy_PP = EmpiricalPP(Toy_lattice, Toy_pff, Toy_energy_cutoff, Toy_atomic_positions,
-                     Toy_nvalence_electrons)
+                     Toy_nvalence_electrons, material="Toy")
 
 # Free electron Pseudopotential ####
 free_lat_centering = "prim"
@@ -748,7 +682,7 @@ free_energy_cutoff = 2*(2*np.pi/free_lat_const)**2
 free_atomic_positions = [[0.]*3]
 free_nvalence_electrons = 1
 free_degree = 2
-free_PP = FreeElectronModel(free_lattice, free_nvalence_electrons, 2)
+free_PP = FreeElectronModel(free_lattice, free_degree)
 
 # The following pseudopotentials come from: 
 # Cohen, Marvin L., and Volker Heine. "The fitting of pseudopotentials to
@@ -767,7 +701,7 @@ Al_energy_cutoff = 4*(2*np.pi/Al_lat_const)**2
 Al_atomic_positions = [[0.,0.,0.]]
 Al_nvalence_electrons = 3
 Al_PP = EmpiricalPP(Al_lattice, Al_pff, Al_energy_cutoff, Al_atomic_positions,
-                    Al_nvalence_electrons)
+                    Al_nvalence_electrons, material="Al")
 
 #### Pseudopotential of Li ####
 Li_centering_type = "body"
@@ -781,7 +715,7 @@ Li_energy_cutoff = 4*(2*np.pi/Li_lat_const)**2
 Li_atomic_positions = [[0.,0.,0.]]
 Li_nvalence_electrons = 1
 Li_PP = EmpiricalPP(Li_lattice, Li_pff, Li_energy_cutoff, Li_atomic_positions,
-                    Li_nvalence_electrons)
+                    Li_nvalence_electrons, material="Li")
 
 #### Pseudopotential of Na ####
 Na_centering_type = "body"
@@ -795,7 +729,7 @@ Na_energy_cutoff = 2*(2*np.pi/Na_lat_const)**2
 Na_atomic_positions = [[0.]*3]
 Na_nvalence_electrons = 1
 Na_PP = EmpiricalPP(Na_lattice, Na_pff, Na_energy_cutoff, Na_atomic_positions,
-                    Na_nvalence_electrons)
+                    Na_nvalence_electrons, material="Na")
     
 #### Pseudopotential of K ####
 K_centering_type = "body"
@@ -809,7 +743,7 @@ K_energy_cutoff = 4*(2*np.pi/K_lat_const)**2
 K_atomic_positions = [[0.]*3]
 K_nvalence_electrons = 1
 K_PP = EmpiricalPP(K_lattice, K_pff, K_energy_cutoff, K_atomic_positions,
-                   K_nvalence_electrons)
+                   K_nvalence_electrons, material="K")
 
 #### Pseudopotential of Rb ####
 Rb_centering_type = "body"
@@ -823,7 +757,7 @@ Rb_energy_cutoff = 2*(2*np.pi/Rb_lat_const)**2
 Rb_atomic_positions = [[0.]*3]
 Rb_nvalence_electrons = 1
 Rb_PP = EmpiricalPP(Rb_lattice, Rb_pff, Rb_energy_cutoff, Rb_atomic_positions,
-                    Rb_nvalence_electrons)
+                    Rb_nvalence_electrons, material="Rb")
 
 #### Pseudopotential of Cs ####
 Cs_centering_type = "body"
@@ -837,7 +771,7 @@ Cs_energy_cutoff = 2*(2*np.pi/Cs_lat_const)**2
 Cs_atomic_positions = [[0.]*3]
 Cs_nvalence_electrons = 1
 Cs_PP = EmpiricalPP(Cs_lattice, Cs_pff, Cs_energy_cutoff, Cs_atomic_positions,
-                    Cs_nvalence_electrons)
+                    Cs_nvalence_electrons, material="Cs")
     
 #### Pseudopotential of Cu ####
 Cu_centering_type = "face"
@@ -851,7 +785,7 @@ Cu_energy_cutoff = 4*(2*np.pi/Cu_lat_const)**2
 Cu_atomic_positions = [[0.]*3]
 Cu_nvalence_electrons = 11
 Cu_PP = EmpiricalPP(Cu_lattice, Cu_pff, Cu_energy_cutoff, Cu_atomic_positions,
-                    Cu_nvalence_electrons)
+                    Cu_nvalence_electrons, material="Cu")
 
 #### Pseudopotential of Ag ####
 Ag_centering_type = "face"
@@ -865,7 +799,7 @@ Ag_energy_cutoff = 4*(2*np.pi/Ag_lat_const)**2
 Ag_atomic_positions = [[0.]*3]
 Ag_nvalence_electrons = 11
 Ag_PP = EmpiricalPP(Ag_lattice, Ag_pff, Ag_energy_cutoff, Ag_atomic_positions,
-                    Ag_nvalence_electrons)
+                    Ag_nvalence_electrons, material="Ag")
 
 #### Pseudopotential of Au ####
 Au_centering_type = "face"
@@ -879,7 +813,7 @@ Au_energy_cutoff = 4*(2*np.pi/Au_lat_const)**2
 Au_atomic_positions = [[0.]*3]
 Au_nvalence_electrons = 11
 Au_PP = EmpiricalPP(Au_lattice, Au_pff, Au_energy_cutoff, Au_atomic_positions,
-                    Au_nvalence_electrons)
+                    Au_nvalence_electrons, material="Au")
 
 #### Pseudopotential of Pb ####
 Pb_centering_type = "face"
@@ -893,7 +827,7 @@ Pb_energy_cutoff = 4*(2*np.pi/Pb_lat_const)**2
 Pb_atomic_positions = [[0.]*3]
 Pb_nvalence_electrons = 4
 Pb_PP = EmpiricalPP(Pb_lattice, Pb_pff, Pb_energy_cutoff, Pb_atomic_positions,
-                    Pb_nvalence_electrons)
+                    Pb_nvalence_electrons, material="Pb")
 
 #### Pseudopotential of Mg ####
 Mg_centering_type = "prim"
@@ -913,7 +847,7 @@ Mg_energy_cutoff = np.dot(Mg_cutoff_vec, Mg_cutoff_vec)
 Mg_atomic_positions = [[0.]*3]
 Mg_nvalence_electrons = 2
 Mg_PP = EmpiricalPP(Mg_lattice, Mg_pff, Mg_energy_cutoff, Mg_atomic_positions,
-                    Mg_nvalence_electrons)
+                    Mg_nvalence_electrons, material="Mg")
 
 #### Pseudopotential of Zn ####
 Zn_centering_type = "prim"
@@ -931,7 +865,7 @@ Zn_energy_cutoff = np.dot(Zn_cutoff_vec, Zn_cutoff_vec)
 Zn_atomic_positions = [[0.]*3]
 Zn_nvalence_electrons = 12
 Zn_PP = EmpiricalPP(Zn_lattice, Zn_pff, Zn_energy_cutoff, Zn_atomic_positions,
-                    Zn_nvalence_electrons)
+                    Zn_nvalence_electrons, material="Zn")
 
 
 # See Band structure and Fermi surface of Zinc and Cadmium by Stark and
@@ -953,7 +887,7 @@ Cd_energy_cutoff = np.dot(Cd_cutoff_vec, Cd_cutoff_vec)
 Cd_atomic_positions = [[0.]*3]
 Cd_nvalence_electrons = 12
 Cd_PP = EmpiricalPP(Cd_lattice, Cd_pff, Cd_energy_cutoff, Cd_atomic_positions,
-                    Cd_nvalence_electrons)
+                    Cd_nvalence_electrons, material="Cd")
 
 #### Pseudopotential of Hg ####
 Hg_centering_type = "prim"
@@ -970,7 +904,7 @@ Hg_energy_cutoff = np.dot(Hg_cutoff_vec, Hg_cutoff_vec)
 Hg_atomic_positions = [[0.]*3]
 Hg_nvalence_electrons = 12
 Hg_PP = EmpiricalPP(Hg_lattice, Hg_pff, Hg_energy_cutoff, Hg_atomic_positions,
-                    Hg_nvalence_electrons)
+                    Hg_nvalence_electrons, material="Hg")
 
 #### Pseudopotential of In ####
 In_centering_type = "body"
@@ -988,7 +922,7 @@ In_energy_cutoff = np.dot(In_cutoff_vec, In_cutoff_vec)
 In_atomic_positions = [[0.]*3]
 In_nvalence_electrons = 3
 In_PP = EmpiricalPP(In_lattice, In_pff, In_energy_cutoff, In_atomic_positions,
-                    In_nvalence_electrons)
+                    In_nvalence_electrons, material="In")
 
 #### Pseudopotential of Sn ####
 Sn_centering_type = "body"
@@ -1006,4 +940,4 @@ Sn_energy_cutoff = np.dot(Sn_cutoff_vec, Sn_cutoff_vec)
 Sn_atomic_positions = [[0.]*3]
 Sn_nvalence_electrons = 4
 Sn_PP = EmpiricalPP(Sn_lattice, Sn_pff, Sn_energy_cutoff, Sn_atomic_positions,
-                    Sn_nvalence_electrons)
+                    Sn_nvalence_electrons, material="Sn")